refactor: move getTransaction/waitForTransaction declarations to IWalletAccountReadOnlySimple
Keep the default waitForTransaction implementation on WalletAccountReadOnly and extract a top-level sleep helper.
