+++ src/wallet-account-read-only.js
@@ -21,6 +21,10 @@ import {
 
 import { IWalletAccountReadOnlySimple } from './wallet-account-read-only-simple.js'
 
+/** @typedef {import('./wallet-account-read-only-simple.js').Finality} Finality */
+/** @typedef {import('./wallet-account-read-only-simple.js').TransactionReceipt} TransactionReceipt */
+/** @typedef {import('./wallet-account-read-only-simple.js').WaitForTransactionOptions} WaitForTransactionOptions */
+
 /**
  * @typedef {Object} Transaction
  * @property {string} to - The transaction's recipient.
@@ -47,17 +51,6 @@ import { IWalletAccountReadOnlySimple } from './wallet-account-read-only-simple.
  */
 
 /**
- * A normalized, cross-chain transaction finality level.
- *
- * - `pending`: seen, not settled (mempool / processed / in-flight).
- * - `confirmed`: settled, reversible only under extreme conditions.
- * - `final`: irreversible per the chain's own guarantees.
- * - `dropped`: evicted / replaced, never landed.
- *
- * @typedef {'pending' | 'confirmed' | 'final' | 'dropped'} Finality
- */
-
-/**
  * Enum that assigns a comparable ordinal to each finality level, used to check
  * whether an observed finality satisfies a requested target.
  *
@@ -72,31 +65,14 @@ export const FINALITY = {
 }
 
 /**
- * A normalized, cross-chain transaction receipt. Blockchain modules extend this
- * type with their own native receipt fields (e.g. `confirmations`, the raw
- * transaction and receipt objects, etc.).
+ * Resolves after the given number of milliseconds.
  *
- * @typedef {Object} TransactionReceipt
- * @property {string} hash - The transaction's identifier (hash / signature / lt:hash).
- * @property {Finality} finality - The transaction's finality level.
- * @property {boolean} [success] - The execution's result (not set if the transaction is still pending or it has been dropped).
- * @property {number} [block] - A reference to the including block (block number / height / slot / masterchain seqno).
- * @property {bigint} [fee] - The fee paid, when known.
+ * @param {number} amount - The delay, in milliseconds.
+ * @returns {Promise<void>} A promise that resolves once the delay elapses.
  */
-
-/**
- * The finality level to wait for.
- *
- * @typedef {'confirmed' | 'final'} WaitForTransactionTarget
- */
-
-/**
- * @typedef {Object} WaitForTransactionOptions
- * @property {WaitForTransactionTarget} [target] - The finality target to wait for (default: 'confirmed').
- * @property {number} [timeout] - The total time budget in milliseconds before giving up (default: 60000).
- * @property {number} [interval] - The poll cadence in milliseconds (default: 4000).
- * @property {number} [maxPollErrors] - How many consecutive getTransaction() failures to tolerate before rethrowing (default: 3).
- */
+async function sleep (amount) {
+  await new Promise(resolve => setTimeout(resolve, amount))
+}
 
 /** @interface */
 export class IWalletAccountReadOnly extends IWalletAccountReadOnlySimple {
@@ -118,41 +94,6 @@ export class IWalletAccountReadOnly extends IWalletAccountReadOnlySimple {
    */
   async quoteTransfer (options) {
     throw new NotImplementedError('quoteTransfer(options)')
-  }
-
-  /**
-   * Returns a transaction's receipt.
-   *
-   * @deprecated Use {@link getTransaction} instead, which returns a normalized, finality-based receipt. The native receipt fields remain available on each module's extended return type.
-   * @param {string} hash - The transaction's hash.
-   * @returns {Promise<unknown | null>} The receipt, or null if the transaction has not been included in a block yet.
-   */
-  async getTransactionReceipt (hash) {
-    throw new NotImplementedError('getTransactionReceipt(hash)')
-  }
-
-  /**
-   * Returns a normalized, finality-based receipt for a transaction.
-   *
-   * @param {string} hash - The transaction's identifier (hash / signature / lt:hash).
-   * @returns {Promise<TransactionReceipt>} The normalized receipt.
-   * @throws {ValueError} If the hash is not a valid identifier.
-   * @throws {NoSuchElementError} If no transaction has been found for the given hash.
-   */
-  async getTransaction (hash) {
-    throw new NotImplementedError('getTransaction(hash)')
-  }
-
-  /**
-   * Blocks until a transaction reaches a terminal state (the requested finality target or `dropped`), or times out.
-   *
-   * @param {string} hash - The transaction's identifier.
-   * @param {WaitForTransactionOptions} [options] - The wait options.
-   * @returns {Promise<TransactionReceipt>} The terminal receipt.
-   * @throws {TimeoutError} If the target is not reached before the timeout.
-   */
-  async waitForTransaction (hash, options) {
-    throw new NotImplementedError('waitForTransaction(hash, options)')
   }
 }
 
@@ -353,7 +294,7 @@ export default class WalletAccountReadOnly {
         throw new TimeoutError(`Transaction '${hash}' did not reach '${target}' within the timeout.`)
       }
 
-      await new Promise(resolve => setTimeout(resolve, interval))
+      await sleep(interval)
     }
   }
 }
