+++ types/src/wallet-account-read-only-simple.d.ts
@@ -1,4 +1,62 @@
 /**
+ * A normalized, cross-chain transaction finality level.
+ *
+ * - `pending`: seen, not settled (mempool / processed / in-flight).
+ * - `confirmed`: settled, reversible only under extreme conditions.
+ * - `final`: irreversible per the chain's own guarantees.
+ * - `dropped`: evicted / replaced, never landed.
+ */
+export type Finality = "pending" | "confirmed" | "final" | "dropped";
+/**
+ * A normalized, cross-chain transaction receipt. Blockchain modules extend this
+ * type with their own native receipt fields (e.g. `confirmations`, the raw
+ * transaction and receipt objects, etc.).
+ */
+export type TransactionReceipt = {
+    /**
+     * - The transaction's identifier (hash / signature / lt:hash).
+     */
+    hash: string;
+    /**
+     * - The transaction's finality level.
+     */
+    finality: Finality;
+    /**
+     * - The execution's result (not set if the transaction is still pending or it has been dropped).
+     */
+    success?: boolean;
+    /**
+     * - A reference to the including block (block number / height / slot / masterchain seqno).
+     */
+    block?: number;
+    /**
+     * - The fee paid, when known.
+     */
+    fee?: bigint;
+};
+/**
+ * The finality level to wait for.
+ */
+export type WaitForTransactionTarget = "confirmed" | "final";
+export type WaitForTransactionOptions = {
+    /**
+     * - The finality target to wait for (default: 'confirmed').
+     */
+    target?: WaitForTransactionTarget;
+    /**
+     * - The total time budget in milliseconds before giving up (default: 60000).
+     */
+    timeout?: number;
+    /**
+     * - The poll cadence in milliseconds (default: 4000).
+     */
+    interval?: number;
+    /**
+     * - How many consecutive getTransaction() failures to tolerate before rethrowing (default: 3).
+     */
+    maxPollErrors?: number;
+};
+/**
  * The read-only members shared by every wallet account, single-signer or multisig.
  *
  * This is an internal base interface: it is not exported from the package entry point.
@@ -39,8 +97,27 @@ export interface IWalletAccountReadOnlySimple {
     /**
      * Returns a transaction's receipt.
      *
+     * @deprecated Use {@link getTransaction} instead, which returns a normalized, finality-based receipt. The native receipt fields remain available on each module's extended return type.
      * @param {string} hash - The transaction's hash.
      * @returns {Promise<unknown | null>} The receipt, or null if the transaction has not been included in a block yet.
      */
     getTransactionReceipt(hash: string): Promise<unknown | null>;
+    /**
+     * Returns a normalized, finality-based receipt for a transaction.
+     *
+     * @param {string} hash - The transaction's identifier (hash / signature / lt:hash).
+     * @returns {Promise<TransactionReceipt>} The normalized receipt.
+     * @throws {ValueError} If the hash is not a valid identifier.
+     * @throws {NoSuchElementError} If no transaction has been found for the given hash.
+     */
+    getTransaction(hash: string): Promise<TransactionReceipt>;
+    /**
+     * Blocks until a transaction reaches a terminal state (the requested finality target or `dropped`), or times out.
+     *
+     * @param {string} hash - The transaction's identifier.
+     * @param {WaitForTransactionOptions} [options] - The wait options.
+     * @returns {Promise<TransactionReceipt>} The terminal receipt.
+     * @throws {TimeoutError} If the target is not reached before the timeout.
+     */
+    waitForTransaction(hash: string, options?: WaitForTransactionOptions): Promise<TransactionReceipt>;
 }
